refactor(ui): isolate composer dictation orchestration
